requirements.txt
pyyaml>=6.0
python-frontmatter>=1.0
google-genai>=0.3.0
python-dotenv>=1.0
lancedb>=0.6.0
pyarrow>=15.0
